Add Github information to Kabanero document for CLI (#170)
* Allow the Github organization to be specified on the Kabanero object
This allows the CLI to query the right information in Github, for the
organization this instance is using.  I refactored the code in the
landing page that programmatically creates a Deployment since the
CLI now needs to do the same.

* Refactoring for github org/team/apiurl and AES encryption key
After discussions with the CLI team, there are more pieces of information
that need to be consumed by the CLI.  The Github org/team are required
to do role mapping.  There is a perceived future need for the github
API URL, although that is optional right now.  The operator will
generate the AES encryption key, which allows JWTs generated by the
CLI microservice to be re-usable across instances of the CLI pod.

* Dave's code review comments

* Inadvertent removals

* Correct cli provisioning for versioning changes