Add SkipCertValidation to the collection CRD
This is the first step in removing the requirement that a Collection
CR instance has a Kabanero CR instance owner.  In the future, an
administrator may choose to create a one-off Collection CR instance
that has no ties to a Kabanero CR instance.  Additional changes will
be required in other components (for example, the CLI service) before
the collection controller can remove its dependency on the Kabanero
CR instance.